feat: implement many changes related to calibration, marker detection, homographies and touch detection
- Updated CalibrationResult class to handle separate depth and RGB homography matrices.
- Added validation for depth and RGB corners in CalibrationResult.
- Introduced CoordinateTransformer for bidirectional coordinate transformations between camera and projector.
- Implemented ImageTransformer for image transformations using the homography matrices.
- Modified TouchDetector to utilize RGB corners for defining the region of interest.
- Enhanced main application logic to integrate image transformation and display.

+++ code/cv-system/src/cv_system/calibration/homography.py
@@ -9,8 +9,8 @@ import numpy as np
 
 
 def compute_homography(
-    camera_points: list[tuple[float, float]],
-    projector_points: list[tuple[float, float]],
+    camera_points: list[tuple[int, int]],
+    projector_points: list[tuple[int, int]],
 ) -> np.ndarray:
     """Compute 3x3 homography matrix from 4 point correspondences.
